upstream

+++ qb/qb.libs.sh
@@ -36,10 +36,12 @@ add_dirs()
 check_compiler()
 {	if [ "$1" = cxx ]; then
 		COMPILER="$CXX"
+		FLAGS="$CXXFLAGS"
 		TEMP_CODE="$TEMP_CXX"
 		TEST_C="extern \"C\" { void $2(void); } int main() { $2(); }"
 	else
 		COMPILER="$CC"
+		FLAGS="$CFLAGS"
 		TEMP_CODE="$TEMP_C"
 		TEST_C="void $2(void); int main(void) { $2(); return 0; }"
 	fi
@@ -85,7 +87,8 @@ check_enabled()
 # $4 = function in lib [checked only if non-empty]
 # $5 = extralibs [checked only if non-empty]
 # $6 = headers [checked only if non-empty]
-# $7 = critical error message [checked only if non-empty]
+# $7 = include directory [checked only if non-empty]
+# $8 = critical error message [checked only if non-empty]
 check_lib()
 {	tmpval="$(eval "printf %s \"\$HAVE_$2\"")"
 	[ "$tmpval" = 'no' ] && return 0
@@ -93,32 +96,50 @@ check_lib()
 	check_compiler "$1" "$4"
 
 	if [ "$4" ]; then
-		ECHOBUF="Checking function $4 in ${3% }"
+		MSG="Checking function $4 in"
 		if [ "$6" ]; then
 			printf %s\\n "$6" "int main(void) { void *p = (void*)$4; return 0; }" > "$TEMP_CODE"
 		else
 			printf %s\\n "$TEST_C" > "$TEMP_CODE"
 		fi
 	else
-		ECHOBUF="Checking existence of ${3% }"
+		MSG='Checking existence of'
 		printf %s\\n 'int main(void) { return 0; }' > "$TEMP_CODE"
 	fi
 
 	val="$2"
-	lib="$3"
-	error="${7:-}"
+	lib="${3% }"
+	include="${7:-}"
+	error="${8:-}"
 	answer='no'
-	eval "set -- $INCLUDE_DIRS $LIBRARY_DIRS $5 $CFLAGS $LDFLAGS $3"
+	printf %s "$MSG $lib"
+	eval "set -- $INCLUDE_DIRS $LIBRARY_DIRS $5 $FLAGS $LDFLAGS $lib"
 	"$COMPILER" -o "$TEMP_EXE" "$TEMP_CODE" "$@" >>config.log 2>&1 && answer='yes'
+	printf %s\\n " ... $answer"
+
+	if [ "$answer" = 'yes' ] && [ "$include" ]; then
+		answer='no'
+		eval "set -- $INCLUDES"
+		for dir do
+			[ "$answer" = 'yes' ] && break
+			printf %s "Checking existence of /$dir/$include"
+			if [ -d "/$dir/$include" ]; then
+				eval "${val}_CFLAGS=\"-I/$dir/$include\""
+				answer='yes'
+			fi
+			printf %s\\n " ... $answer"
+		done
+	fi
+
 	eval "HAVE_$val=\"$answer\""
-	printf %s\\n "$ECHOBUF ... $answer"
 	rm -f -- "$TEMP_CODE" "$TEMP_EXE"
 
 	if [ "$answer" = 'no' ]; then
 		[ "$error" ] && die 1 "$error"
-		[ "$tmpval" = 'yes' ] && {
+		setval="$(eval "printf %s \"\$USER_$val\"")"
+		if [ "$setval" = 'yes' ]; then
 			die 1 "Forced to build with library $lib, but cannot locate. Exiting ..."
-		}
+		fi
 	else
 		eval "${val}_LIBS=\"$lib\""
 		PKG_CONF_USED="$PKG_CONF_USED $val"
@@ -130,48 +151,62 @@ check_lib()
 # check_pkgconf:
 # If available uses $PKG_CONF_PATH to find a library.
 # $1 = HAVE_$1
-# $2 = package
+# $2 = package ['package' or 'package package1 package2', $1 = name]
 # $3 = version [checked only if non-empty]
 # $4 = critical error message [checked only if non-empty]
+# $5 = force check_lib when true [checked only if non-empty, set by check_val]
 check_pkgconf()
 {	tmpval="$(eval "printf %s \"\$HAVE_$1\"")"
 	eval "TMP_$1=\$tmpval"
 	[ "$tmpval" = 'no' ] && return 0
 
-	ECHOBUF="Checking presence of package $2"
-	[ "$3" ] && ECHOBUF="$ECHOBUF >= ${3##* }"
+	ECHOBUF=''
+	[ "${3:-}" ] && ECHOBUF=" >= ${3##* }"
+
+	pkg="${2%% *}"
+	MSG='Checking presence of package'
 
 	[ "$PKG_CONF_PATH" = "none" ] && {
 		eval "HAVE_$1=no"
 		eval "${1#HAVE_}_VERSION=0.0"
-		printf %s\\n "$ECHOBUF ... no"
+		printf %s\\n "$MSG $pkg$ECHOBUF ... no"
 		return 0
 	}
 
 	val="$1"
-	pkg="$2"
-	err="$4"
+	ver="${3:-0.0}"
+	err="${4:-}"
+	lib="${5:-}"
 	answer='no'
 	version='no'
 
-	eval "set -- ${3:-0.0}"
-	for ver do
-		if $PKG_CONF_PATH --atleast-version="$ver" "$pkg"; then
-			answer='yes'
-			version="$("$PKG_CONF_PATH" --modversion "$pkg")"
-			eval "${val}_CFLAGS=\"$("$PKG_CONF_PATH" "$pkg" --cflags)\""
-			eval "${val}_LIBS=\"$("$PKG_CONF_PATH" "$pkg" --libs)\""
-			eval "${val#HAVE_}_VERSION=\"$ver\""
-			break
-		fi
+	eval "set -- ${2#* }"
+	for pkgnam do
+		[ "$answer" = 'yes' ] && break
+		printf %s "$MSG $pkgnam$ECHOBUF"
+		eval "set -- $ver"
+		for pkgver do
+			if "$PKG_CONF_PATH" --atleast-version="$pkgver" "$pkgnam"; then
+				answer='yes'
+				version="$("$PKG_CONF_PATH" --modversion "$pkgnam")"
+				eval "${val}_CFLAGS=\"$("$PKG_CONF_PATH" --cflags "$pkgnam")\""
+				eval "${val}_LIBS=\"$("$PKG_CONF_PATH" --libs "$pkgnam")\""
+				eval "${val}_VERSION=\"$pkgver\""
+				break
+			fi
+		done
+		printf %s\\n " ... $version"
 	done
 
 	eval "HAVE_$val=\"$answer\""
-	printf %s\\n "$ECHOBUF ... $version"
+
 	if [ "$answer" = 'no' ]; then
+		[ "$lib" != 'true' ] || return 0
 		[ "$err" ] && die 1 "$err"
-		[ "$tmpval" = 'yes' ] &&
+		setval="$(eval "printf %s \"\$USER_$val\"")"
+		if [ "$setval" = 'yes' ]; then
 			die 1 "Forced to build with package $pkg, but cannot locate. Exiting ..."
+		fi
 	else
 		PKG_CONF_USED="$PKG_CONF_USED $val"
 	fi
@@ -193,13 +228,16 @@ check_header()
 	done
 	printf %s\\n "int main(void) { return 0; }" >> "$TEMP_C"
 	answer='no'
+	printf %s "Checking presence of header file $CHECKHEADER"
 	eval "set -- $INCLUDE_DIRS"
 	"$CC" -o "$TEMP_EXE" "$TEMP_C" "$@" >>config.log 2>&1 && answer='yes'
 	eval "HAVE_$val=\"$answer\""
-	printf %s\\n "Checking presence of header file $CHECKHEADER ... $answer"
+	printf %s\\n " ... $answer"
 	rm -f -- "$TEMP_C" "$TEMP_EXE"
-	[ "$tmpval" = 'yes' ] && [ "$answer" = 'no' ] && \
+	setval="$(eval "printf %s \"\$USER_$val\"")"
+	if [ "$setval" = 'yes' ] && [ "$answer" = 'no' ]; then
 		die 1 "Build assumed that $header exists, but cannot locate. Exiting ..."
+	fi
 }
 
 # check_macro:
@@ -209,12 +247,11 @@ check_header()
 check_macro()
 {	tmpval="$(eval "printf %s \"\$HAVE_$1\"")"
 	[ "$tmpval" = 'no' ] && return 0
-	if [ "${3}" ]; then
-		ECHOBUF="Checking presence of predefined macro $2 in $3"
+	header_include=''
+	ECHOBUF=''
+	if [ "${3:-}" ]; then
 		header_include="#include <$3>"
-	else
-		ECHOBUF="Checking presence of predefined macro $2"
-		header_include=""
+		ECHOBUF=" in $3"
 	fi
 	cat << EOF > "$TEMP_C"
 $header_include
@@ -226,13 +263,16 @@ EOF
 	answer='no'
 	val="$1"
 	macro="$2"
+	printf %s "Checking presence of predefined macro $macro$ECHOBUF"
 	eval "set -- $CFLAGS $INCLUDE_DIRS"
 	"$CC" -o "$TEMP_EXE" "$TEMP_C" "$@" >>config.log 2>&1 && answer='yes'
 	eval "HAVE_$val=\"$answer\""
-	printf %s\\n "$ECHOBUF ... $answer"
+	printf %s\\n " ... $answer"
 	rm -f -- "$TEMP_C" "$TEMP_EXE"
-	[ "$tmpval" = 'yes' ] && [ "$answer" = 'no' ] && \
+	setval="$(eval "printf %s \"\$USER_$val\"")"
+	if [ "$setval" = 'yes' ] && [ "$answer" = 'no' ]; then
 		die 1 "Build assumed that $macro is defined, but it's not. Exiting ..."
+	fi
 }
 
 # check_switch:
@@ -243,16 +283,16 @@ EOF
 check_switch()
 {	check_compiler "$1" ''
 
-	ECHOBUF="Checking for availability of switch $3 in $COMPILER"
 	printf %s\\n 'int main(void) { return 0; }' > "$TEMP_CODE"
 	answer='no'
+	printf %s "Checking for availability of switch $3 in $COMPILER"
 	"$COMPILER" -o "$TEMP_EXE" "$TEMP_CODE" "$3" >>config.log 2>&1 && answer='yes'
 	eval "HAVE_$2=\"$answer\""
-	printf %s\\n "$ECHOBUF ... $answer"
+	printf %s\\n " ... $answer"
 	rm -f -- "$TEMP_CODE" "$TEMP_EXE"
-	[ "$answer" = 'no' ] && {
-		[ "$4" ] && die 1 "$4"
-	}
+	if [ "$answer" = 'no' ] && [ "${4:-}" ]; then
+		die 1 "$4"
+	fi
 }
 
 # check_val:
@@ -266,27 +306,13 @@ check_switch()
 # $7 = critical error message [checked only if non-empty]
 # $8 = force check_lib when true [checked only if non-empty]
 check_val()
-{	check_pkgconf "$2" "$5" "${6:-}" "${7:-}"
+{	check_pkgconf "$2" "$5" "${6:-}" "${7:-}" "${8:-}"
 	[ "$PKG_CONF_PATH" = "none" ] || [ "${8:-}" = true ] || return 0
 	tmpval="$(eval "printf %s \"\$HAVE_$2\"")"
 	oldval="$(eval "printf %s \"\$TMP_$2\"")"
 	if [ "$tmpval" = 'no' ] && [ "$oldval" != 'no' ]; then
 		eval "HAVE_$2=auto"
-		check_lib "$1" "$2" "$3"
-
-		if [ "${4:-}" ] && [ "$answer" = 'yes' ]; then
-			val="$2"
-			include="$4"
-			eval "set -- $INCLUDES"
-			for dir do
-				[ -d "/$dir/$include" ] && { eval "${val}_CFLAGS=\"-I/$dir/$include\""; break; }
-			done
-			[ -z "$(eval "printf %s \"\${${val}_CFLAGS}\"")" ] && eval "HAVE_$val=no"
-		fi
-
-		if [ "$answer" = 'no' ] && [ "$oldval" = 'yes' ]; then
-			die 1 "Forced to build with library $lib, but cannot locate. Exiting ..."
-		fi
+		check_lib "$1" "$2" "$3" '' '' '' "${4:-}" "${7:-}"
 	fi
 }
